TP_Solid_Completo

+++ src/ar/edu/unq/po2/tp4/Factura.java
@@ -4,7 +4,7 @@ abstract class Factura implements Pagar, Agencia {
 	
 	double tasaServicio;
 	double costo;
-	double unidadConsumida;
+	int unidadConsumida;
 
 	Factura(double tasaServicio) {
 		this.tasaServicio = tasaServicio;
